add get env commands

+++ shared/lib/with.ps1
@@ -40,19 +40,74 @@ function Invoke-WithEnvironment {
 
 function Set-EnvironmentVariable {
   param(
-    [Parameter(Mandatory=$true, Position=0)][string]$Name,
-    [Parameter(Mandatory=$true, Position=1)][string]$Value,
-    [Parameter(Mandatory=$false)][System.EnvironmentVariableTarget]$Scope='Process'
+    [Parameter(Mandatory=$true,  Position=0)][string]$Name,
+    [Parameter(Mandatory=$true,  Position=1, ValueFromPipeline=$true)]
+    [string]$Value,
+    [Parameter(Mandatory=$false, Position=2)][Alias('s')]
+    [System.EnvironmentVariableTarget]$Scope='Process',
+    [Parameter(Mandatory=$false,
+      HelpMessage='Always on if Scope value is the same as current PS value.')]
+    [Alias('p')]
+    [switch]$CopyToPS=$false
   )
-  [System.Environment]::SetEnvironmentVariable($Name, $Value, $Scope)
-  Set-Content -Path Env:\$Name -Value $Value
+  $processSameAsScope = $false
+  if ($Scope -eq 'Process') {
+    $processSameAsScope = $true
+  } else {
+    $scorig = [System.Environment]::GetEnvironmentVariable($Name, $Scope)
+    $psorig = Get-Content -Path Env:\$Name -ErrorAction Ignore
+    if ([string]::IsNullOrEmpty($psorig) -or $scorig -eq $psorig) {
+      $processSameAsScope = $true
+    }
+  }
+  Write-Host -NoNewline "Warning: This call sometimes takes a million years for unknown reasons... "
+  try {
+    [System.Environment]::SetEnvironmentVariable($Name, $Value, $Scope)
+    if ($processSameAsScope -or $CopyToPS) {
+      Set-Content -Path Env:\$Name -Value $Value
+    }
+  } finally {
+    Write-Host "done!"
+  }
+}
+
+function Get-EnvironmentVariable {
+  param(
+    [Parameter(Mandatory=$true,  Position=0)][string]$Name,
+    [Parameter(Mandatory=$false, Position=1)][Alias('s')]
+    [System.EnvironmentVariableTarget]$Scope='Process'
+  )
+  [System.Environment]::GetEnvironmentVariable($Name, $Scope)
+}
+
+function Get-EnvironmentVariables {
+  param(
+    [Parameter(Mandatory=$false, Position=0)][Alias('s')]
+    [System.EnvironmentVariableTarget]$Scope='Process'
+  )
+  [System.Environment]::GetEnvironmentVariables($Scope)
 }
 
 function Remove-EnvironmentVariable {
   param(
     [Parameter(Mandatory=$true, Position=0)][string]$Name,
-    [Parameter(Mandatory=$false)][System.EnvironmentVariableTarget]$Scope='Process'
+    [Parameter(Mandatory=$false)][Alias('s')]
+    [System.EnvironmentVariableTarget]$Scope='Process',
+    [Parameter(Mandatory=$false, HelpMessage='Always on if Scope is Process.')]
+    [Alias('p')][switch]$RemoveFromPS=$false
   )
-  [System.Environment]::SetEnvironmentVariable($Name, '', $Scope)
-  Remove-Item -Pat Env:\$Name
+  $processSameAsScope = $false
+  if ($Scope -eq 'Process') {
+    $processSameAsScope = $true
+  } else {
+    $scorig = [System.Environment]::GetEnvironmentVariable($Name, $Scope)
+    $psorig = Get-Content -Path Env:\$Name -ErrorAction Ignore
+    if ([string]::IsNullOrEmpty($psorig) -or $scorig -eq $psorig) {
+      $processSameAsScope = $true
+    }
+  }
+  [System.Environment]::SetEnvironmentVariable($Name, $null, $Scope)
+  if ($processSameAsScope -or $RemoveFromPS) {
+    Remove-Item -Pat Env:\$Name
+  }
 }
